opt: rename `build` to `optbuilder`
Release note: None
