+++ pkg/sql/opt/exec/execbuilder/builder_test.go
@@ -26,8 +26,8 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/internal/client"
-	optbuild "github.com/cockroachdb/cockroach/pkg/sql/opt/build"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/exec"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/optbuilder"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/xform"
 	"github.com/cockroachdb/cockroach/pkg/sql/optbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
@@ -83,7 +83,7 @@ func TestBuild(t *testing.T) {
 
 					// Build and optimize the opt expression tree.
 					o := xform.NewOptimizer(catalog, xform.OptimizeNone)
-					root, props, err := optbuild.NewBuilder(ctx, o.Factory(), stmt).Build()
+					root, props, err := optbuilder.NewBuilder(ctx, o.Factory(), stmt).Build()
 					if err != nil {
 						d.Fatalf(t, "BuildOpt: %v", err)
 					}
