+++ pkg/sql/opt/optbuilder/builder_test.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package build
+package optbuilder
 
 // This file is home to TestBuilder, which is similar to the logic tests, except it
 // is used for optimizer builder-specific testcases.
