+++ pkg/sql/opt/optbuilder/column_props.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package build
+package optbuilder
 
 import (
 	"fmt"
