+++ pkg/sql/opt/optbuilder/testdata/aggregate
