+++ pkg/sql/opt/optbuilder/testdata/distinct
