+++ pkg/sql/opt/optbuilder/testdata/having
