+++ pkg/sql/opt/optbuilder/testdata/inner-join
