+++ pkg/sql/opt/optbuilder/testdata/scalar
